1.3.0

+++ source/assets/data/sql-init.js
@@ -1,6 +1,6 @@
 export const sqlInit =
     `CREATE TABLE IF NOT EXISTS "user" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "name"	TEXT NOT NULL,
     "username"	TEXT NOT NULL UNIQUE,
     "password"	TEXT NOT NULL,
@@ -10,7 +10,7 @@ export const sqlInit =
 );
 
     CREATE TABLE IF NOT EXISTS "product" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "name"	TEXT NOT NULL,
     "description"	TEXT,
     "category"  TEXT,
@@ -22,7 +22,7 @@ export const sqlInit =
 );
 
     CREATE TABLE IF NOT EXISTS "client" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "name"	TEXT NOT NULL,
     "address"	TEXT,
     "phone"	TEXT,
@@ -34,7 +34,7 @@ export const sqlInit =
 );
 
     CREATE TABLE IF NOT EXISTS "supplier" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "name"	TEXT NOT NULL,
     "address"	TEXT,
     "phone"	TEXT,
@@ -46,7 +46,7 @@ export const sqlInit =
 );
 
     CREATE TABLE IF NOT EXISTS "employee" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "name"	TEXT NOT NULL,
     "address"	TEXT,
     "phone"	TEXT,
@@ -59,13 +59,13 @@ export const sqlInit =
 );
 
     CREATE TABLE IF NOT EXISTS "payment" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "amount"	NUMERIC NOT NULL,
     "date"  TEXT NOT NULL,
     "comment"  TEXT NOT NULL,
     "error"  BOOLEAN,
-    "employee"	INTEGER NOT NULL,
-    "responsible"	INTEGER NOT NULL,
+    "employee"	BIGINT NOT NULL,
+    "responsible"	BIGINT NOT NULL,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(employee) REFERENCES "employee"(id),
     FOREIGN KEY(responsible) REFERENCES "user"(id),
@@ -73,24 +73,24 @@ export const sqlInit =
 );
 
 CREATE TABLE IF NOT EXISTS "register" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "amount"	NUMERIC NOT NULL,
     "date"  TEXT NOT NULL,
     "comment"  TEXT NOT NULL,
     "error"  BOOLEAN,
-    "responsible"	INTEGER NOT NULL,
+    "responsible"	BIGINT NOT NULL,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(responsible) REFERENCES "user"(id),
     PRIMARY KEY("id")
 );
   
     CREATE TABLE IF NOT EXISTS "comment" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "comment"	TEXT NOT NULL,
     "date"  TEXT NOT NULL,
-    "client"	INTEGER,
-    "employee"	INTEGER,
-    "supplier"	INTEGER,
+    "client"	BIGINT,
+    "employee"	BIGINT,
+    "supplier"	BIGINT,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(client) REFERENCES "client"(id),
     FOREIGN KEY(employee) REFERENCES "employee"(id),
@@ -99,7 +99,7 @@ CREATE TABLE IF NOT EXISTS "register" (
 );
 
     CREATE TABLE IF NOT EXISTS "transaction" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "goldIn"	NUMERIC NOT NULL,
     "amountIn"	NUMERIC NOT NULL,
     "goldOut"	NUMERIC NOT NULL,
@@ -110,8 +110,8 @@ CREATE TABLE IF NOT EXISTS "register" (
     "date"  TEXT NOT NULL,
     "comment"  TEXT,
     "type" TEXT,
-    "client"	INTEGER NOT NULL,
-    "responsible"	INTEGER NOT NULL,
+    "client"	BIGINT NOT NULL,
+    "responsible"	BIGINT NOT NULL,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(client) REFERENCES "client"(id),
     FOREIGN KEY(responsible) REFERENCES "user"(id),
@@ -119,7 +119,7 @@ CREATE TABLE IF NOT EXISTS "register" (
 );
 
 CREATE TABLE IF NOT EXISTS "buy" (
-    "id"	SERIAL NOT NULL,
+    "id"	BIGSERIAL NOT NULL,
     "goldIn"	NUMERIC NOT NULL,
     "amountIn"	NUMERIC NOT NULL,
     "goldOut"	NUMERIC NOT NULL,
@@ -130,8 +130,8 @@ CREATE TABLE IF NOT EXISTS "buy" (
     "date"  TEXT NOT NULL,
     "comment"  TEXT,
     "type" TEXT,
-    "supplier"	INTEGER NOT NULL,
-    "responsible"	INTEGER NOT NULL,
+    "supplier"	BIGINT NOT NULL,
+    "responsible"	BIGINT NOT NULL,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(supplier) REFERENCES "supplier"(id),
     FOREIGN KEY(responsible) REFERENCES "user"(id),
@@ -139,9 +139,9 @@ CREATE TABLE IF NOT EXISTS "buy" (
 );
 
 CREATE TABLE IF NOT EXISTS "transactionProduct" (
-    "id" SERIAL NOT NULL,
-    "product" INTEGER NOT NULL,
-    "transaction" INTEGER NOT NULL,
+    "id" BIGSERIAL NOT NULL,
+    "product" BIGINT NOT NULL,
+    "transaction" BIGINT NOT NULL,
     "quantity"  NUMERIC NOT NULL,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(transaction) REFERENCES "transaction"(id),
@@ -150,9 +150,9 @@ CREATE TABLE IF NOT EXISTS "transactionProduct" (
 );
 
 CREATE TABLE IF NOT EXISTS "buyProduct" (
-    "id" SERIAL NOT NULL,
-    "product" INTEGER NOT NULL,
-    "buy" INTEGER NOT NULL,
+    "id" BIGSERIAL NOT NULL,
+    "product" BIGINT NOT NULL,
+    "buy" BIGINT NOT NULL,
     "quantity"  NUMERIC NOT NULL,
     "deleted" BOOLEAN DEFAULT false,
     FOREIGN KEY(buy) REFERENCES "buy"(id),
